# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MyMemoryCard API",
    "version": "v1"
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PlatformListResponse": {
        "type": "object",
        "properties": {
          "platforms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Platform"
            },
            "maxItems": 500
          }
        },
        "required": [
          "platforms"
        ]
      },
      "Platform": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Platform unique identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "igdb_platform_id": {
            "type": "integer",
            "nullable": true,
            "description": "IGDB platform identifier",
            "example": 6
          },
          "name": {
            "type": "string",
            "description": "Platform name",
            "example": "PC (Windows)"
          },
          "abbreviation": {
            "type": "string",
            "nullable": true,
            "description": "Platform abbreviation",
            "example": "PC"
          },
          "slug": {
            "type": "string",
            "nullable": true,
            "description": "URL-friendly platform identifier",
            "example": "win"
          },
          "platform_family": {
            "type": "string",
            "nullable": true,
            "description": "Platform family grouping",
            "example": "PC"
          },
          "color_primary": {
            "type": "string",
            "description": "Brand color in hex format",
            "example": "#6B7280"
          },
          "created_at": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "Creation timestamp"
          }
        },
        "required": [
          "id",
          "igdb_platform_id",
          "name",
          "abbreviation",
          "slug",
          "platform_family",
          "color_primary",
          "created_at"
        ]
      },
...
      "UserPlatformsListResponse": {
        "type": "object",
        "properties": {
          "user_platforms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPlatformResponse"
            }
          }
        },
        "required": [
          "user_platforms"
        ],
        "description": "List of user platforms"
      },
      "UserPlatformResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "platform_id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "icon_url": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "profile_url": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "user_id",
          "platform_id",
          "created_at"
        ],
        "description": "User-platform association response",
        "example": {
          "id": "123e4567-e89b-12d3-a456-426614174000",
          "user_id": "user-uuid",
          "platform_id": "platform-uuid",
          "username": "my_username",
          "icon_url": "https://example.com/icon.png",
          "profile_url": "https://example.com/profile",
          "notes": "My primary gaming platform",
          "created_at": "2026-01-14T10:00:00Z"
        }
      },
      "AddUserPlatformRequest": {
        "type": "object",
        "properties": {
          "platform_id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "minLength": 1
          },
          "icon_url": {
            "type": "string",
            "format": "uri"
          },
          "profile_url": {
            "type": "string",
            "format": "uri"
          },
          "notes": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "platform_id"
        ],
        "additionalProperties": false,
        "description": "Request to add a platform to user account",
        "example": {
          "platform_id": "123e4567-e89b-12d3-a456-426614174000",
          "username": "my_username",
          "icon_url": "https://example.com/icon.png",
          "profile_url": "https://example.com/profile",
          "notes": "My primary gaming platform"
        }
      },
...
      "GetPreferencesResponse": {
        "type": "object",
        "properties": {
          "preferences": {
            "$ref": "#/components/schemas/PreferencesResponse"
          }
        },
        "required": [
          "preferences"
        ],
        "description": "Get preferences response"
      },
      "PreferencesResponse": {
        "type": "object",
        "properties": {
          "default_view": {
            "type": "string",
            "enum": [
              "grid",
              "table"
            ]
          },
          "items_per_page": {
            "type": "integer",
            "enum": [
              10,
              25,
              50,
              100
            ]
          },
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "auto"
            ]
          },
          "updated_at": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "default_view",
          "items_per_page",
          "theme",
          "updated_at"
        ],
        "description": "User preferences",
        "example": {
          "default_view": "grid",
          "items_per_page": 25,
          "theme": "dark",
          "updated_at": "2026-01-15T10:00:00.000Z"
        }
      },
      "UpdatePreferencesRequest": {
        "type": "object",
        "properties": {
          "default_view": {
            "type": "string",
            "enum": [
              "grid",
              "table"
            ]
          },
          "items_per_page": {
            "type": "integer",
            "enum": [
              10,
              25,
              50,
              100
            ]
          },
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "auto"
            ]
          }
        },
        "additionalProperties": false,
        "description": "Request to update user preferences",
        "example": {
          "default_view": "table",
          "items_per_page": 50,
          "theme": "dark"
        }
      },
...
      "CredentialListResponse": {
        "type": "object",
        "properties": {
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CredentialStatus"
            },
            "maxItems": 10
          }
        },
        "required": [
          "services"
        ],
        "description": "List of all credential statuses for the user"
      },
...
      "CredentialSaveResponse": {
        "type": "object",
        "properties": {
          "service": {
            "type": "string",
            "enum": [
              "igdb",
              "steam",
              "retroachievements",
              "rawg"
            ]
          },
          "credential_type": {
            "type": "string",
            "enum": [
              "twitch_oauth",
              "steam_openid",
              "api_key"
            ]
          },
          "is_active": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "service",
          "credential_type",
          "is_active",
          "message"
        ],
        "description": "Response after saving credentials",
        "example": {
          "service": "igdb",
          "credential_type": "twitch_oauth",
          "is_active": true,
          "message": "Credentials saved for igdb. Please validate to confirm they work."
        }
      },
      "SaveCredentialRequest": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "service": {
                "type": "string",
                "enum": [
                  "igdb",
                  "steam",
                  "retroachievements",
                  "rawg"
                ]
              }
            },
            "required": [
              "service"
            ]
          },
          {
            "oneOf": [
              {
                "type": "object",
                "properties": {
                  "credential_type": {
                    "type": "string",
                    "enum": [
                      "twitch_oauth"
                    ]
                  },
                  "credentials": {
                    "type": "object",
                    "properties": {
                      "client_id": {
                        "type": "string",
                        "minLength": 1
                      },
                      "client_secret": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "required": [
                      "client_id",
                      "client_secret"
                    ]
                  }
                },
                "required": [
                  "credential_type",
                  "credentials"
                ]
              },
              {
                "type": "object",
                "properties": {
                  "credential_type": {
                    "type": "string",
                    "enum": [
                      "api_key"
                    ]
                  },
                  "credentials": {
                    "type": "object",
                    "properties": {
                      "username": {
                        "type": "string"
                      },
                      "api_key": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "required": [
                      "api_key"
                    ]
                  }
                },
                "required": [
                  "credential_type",
                  "credentials"
                ]
              },
              {
                "type": "object",
                "properties": {
                  "credential_type": {
                    "type": "string",
                    "enum": [
                      "steam_openid"
                    ]
                  },
                  "credentials": {
                    "type": "object",
                    "properties": {
                      "steam_id": {
                        "type": "string",
                        "minLength": 1
                      },
                      "display_name": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "steam_id"
                    ]
                  }
                },
                "required": [
                  "credential_type",
                  "credentials"
                ]
              }
            ]
          }
        ],
        "description": "Request to save API credentials for a service",
        "example": {
          "service": "igdb",
          "credential_type": "twitch_oauth",
          "credentials": {
            "client_id": "REPLACE_WITH_CLIENT_ID",
            "client_secret": "REPLACE_WITH_CLIENT_SECRET"
          }
        }
      },
...
      "UserGameCreateRequest": {
        "type": "object",
        "properties": {
          "game_id": {
            "type": "string",
            "format": "uuid"
          },
          "platform_id": {
            "type": "string",
            "format": "uuid"
          },
          "store_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "game_id",
          "platform_id"
        ],
        "additionalProperties": false,
        "description": "Request to add a game to user library on a platform",
        "example": {
          "game_id": "550e8400-e29b-41d4-a716-446655440010",
          "platform_id": "550e8400-e29b-41d4-a716-446655440000",
          "store_id": "550e8400-e29b-41d4-a716-446655440001"
        }
      },
      "UserGameListResponse": {
        "type": "object",
        "properties": {
          "user_games": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserGameResponse"
            },
            "maxItems": 500
          }
        },
        "required": [
          "user_games"
        ],
        "description": "List of user game library entries"
      },
...
      "StoreListResponse": {
        "type": "object",
        "properties": {
          "stores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Store"
            },
            "maxItems": 100
          }
        },
        "required": [
          "stores"
        ]
      },
...
      "RAValidateResponse": {
        "type": "object",
        "properties": {
          "is_valid": {
            "type": "boolean",
            "description": "Whether credentials are valid"
          },
          "message": {
            "type": "string",
            "description": "Validation message"
          }
        },
        "required": [
          "is_valid"
        ],
        "description": "RetroAchievements credentials validation result"
      },
      "RACredentialsRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "description": "RetroAchievements username",
            "example": "YourUsername"
          },
          "api_key": {
            "type": "string",
            "minLength": 1,
            "description": "RetroAchievements API key",
            "example": "your-api-key-here"
          }
        },
        "required": [
          "username",
          "api_key"
        ],
        "description": "RetroAchievements credentials for validation"
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/platforms": {
      "get": {
        "tags": [
          "platforms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List platforms",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/user-platforms": {
      "get": {
        "tags": [
          "user-platforms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user platforms",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPlatformsListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "user-platforms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUserPlatformRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Platform added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPlatformResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/preferences": {
      "get": {
        "tags": [
          "preferences"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User preferences retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPreferencesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "preferences"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preferences updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPreferencesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/credentials": {
      "get": {
        "tags": [
          "credentials"
        ],
        "summary": "List credential statuses",
        "description": "Get the status of all stored API credentials for the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of credential statuses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "credentials"
        ],
        "summary": "Save API credentials",
        "description": "Save or update API credentials for a service. Credentials are encrypted before storage.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveCredentialRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Credentials saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialSaveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/user-games": {
      "post": {
        "tags": [
          "user-games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserGameCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game added to user library successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGameResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Platform or store not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - user already owns this game on this platform",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "user-games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0,
              "default": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "User games retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGameListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stores": {
      "get": {
        "tags": [
          "stores"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List stores",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/retroachievements/credentials": {
      "post": {
        "tags": [
          "retroachievements"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RACredentialsRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Credentials saved and validated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAValidateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}